More snakemake refactoring
