+++ tests/snakemake/test_snakemake_anib_workflow.py
@@ -34,9 +34,12 @@ from pathlib import Path
 import pytest
 
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
-from pyani_plus.snakemake import snakemake_scheduler
-from pyani_plus.tools import ToolExecutor, get_blastn, get_makeblastdb
-from pyani_plus.workflows import check_input_stems
+from pyani_plus.tools import get_blastn, get_makeblastdb
+from pyani_plus.workflows import (
+    ToolExecutor,
+    check_input_stems,
+    run_snakemake_with_progress_bar,
+)
 
 from . import compare_matrices
 
@@ -75,12 +78,13 @@ def test_snakemake_rule_fragments(
     # Remove the output directory to force re-running the snakemake rule
     shutil.rmtree(anib_targets_outdir, ignore_errors=True)
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_anib.smk"
-    )
-    runner.run_workflow(
-        anib_targets_fragments, config_anib_args, workdir=Path(tmp_path)
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_anib.smk",
+        targets=anib_targets_fragments,
+        params=config_anib_args,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
 
     # Check output against target fixtures
@@ -114,11 +118,14 @@ def test_snakemake_rule_blastdb(
     # Remove the output directory to force re-running the snakemake rule
     shutil.rmtree(anib_targets_outdir, ignore_errors=True)
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_anib.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_anib.smk",
+        targets=anib_targets_blastdb,
+        params=config_anib_args,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(anib_targets_blastdb, config_anib_args, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in anib_targets_blastdb:
@@ -160,11 +167,14 @@ def test_snakemake_rule_blastn(  # noqa: PLR0913
     )
     assert db.is_file()
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_anib.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_anib.smk",
+        targets=anib_targets_blastn,
+        params=config_anib_args,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(anib_targets_blastn, config_anib_args, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in anib_targets_blastn:
