+++ tests/snakemake/test_snakemake_dnadiff_workflow.py
@@ -32,15 +32,17 @@ from pathlib import Path
 import pytest
 
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
-from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import (
-    ToolExecutor,
     get_delta_filter,
     get_nucmer,
     get_show_coords,
     get_show_diff,
 )
-from pyani_plus.workflows import check_input_stems
+from pyani_plus.workflows import (
+    ToolExecutor,
+    check_input_stems,
+    run_snakemake_with_progress_bar,
+)
 
 from . import compare_matrices
 
@@ -124,11 +126,14 @@ def test_snakemake_rule_delta(
     config = config_dnadiff_args.copy()
     config["outdir"] = dnadiff_nucmer_targets_delta_outdir
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_dnadiff.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_dnadiff.smk",
+        targets=dnadiff_nucmer_targets_delta,
+        params=config,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(dnadiff_nucmer_targets_delta, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in dnadiff_nucmer_targets_delta:
@@ -162,11 +167,14 @@ def test_snakemake_rule_filter(
     config = config_dnadiff_args.copy()
     config["outdir"] = dnadiff_nucmer_targets_filter_outdir
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_dnadiff.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_dnadiff.smk",
+        targets=dnadiff_nucmer_targets_filter,
+        params=config,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(dnadiff_nucmer_targets_filter, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in dnadiff_nucmer_targets_filter:
@@ -225,11 +233,14 @@ def test_snakemake_rule_show_diff_and_coords(  # noqa: PLR0913
     config = config_dnadiff_args.copy()
     config["outdir"] = dnadiff_targets_showdiff_outdir
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_dnadiff.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_dnadiff.smk",
+        targets=dnadiff_targets_showdiff,
+        params=config,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(dnadiff_targets_showdiff, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in dnadiff_targets_showdiff:
@@ -305,11 +316,14 @@ def test_snakemake_rule_show_coords(  # noqa: PLR0913
     config = config_dnadiff_args.copy()
     config["outdir"] = dnadiff_targets_showcoords_outdir
 
-    # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner(
-        ToolExecutor.local, "snakemake_dnadiff.smk"
+    run_snakemake_with_progress_bar(
+        executor=ToolExecutor.local,
+        workflow_name="snakemake_dnadiff.smk",
+        targets=dnadiff_targets_showcoords,
+        params=config,
+        working_directory=Path(tmp_path),
+        show_progress_bar=False,
     )
-    runner.run_workflow(dnadiff_targets_showcoords, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
     for fname in dnadiff_targets_showcoords:
